Push api

+++ back-end/src/main/java/com/dangki/data/repository/UserRepository.java
@@ -1,8 +1,8 @@
 package com.dangki.data.repository;
 
-import com.dangki.data.entites.User;
+import com.dangki.data.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     User findByUsername(String username);
 }
